Fix the possible branches sync failure when user push branches directly but not visit UI after upgrading from v1.20 -> v1.21
